Moved some stuff around, now the player states are easier to find and edit
