+++ .idea/.idea.Focuslash/.idea/contentModel.xml
@@ -12,17 +12,20 @@
           </e>
         </e>
         <e p="Scripts" t="Include">
-          <e p="Creatures" t="Include">
-            <e p="Creature.cs" t="Include" />
-            <e p="Enemy" t="Include">
-              <e p="BigKnight.cs" t="Include" />
-              <e p="EnemyAttack.cs" t="Include" />
-              <e p="EnemySpawner.cs" t="Include" />
-              <e p="GruntScript.cs" t="Include" />
-              <e p="LaserTurret.cs" t="Include" />
-            </e>
+          <e p="Creature.cs" t="Include" />
+          <e p="Enemy" t="Include">
+            <e p="BigKnight.cs" t="Include" />
+            <e p="EnemyAttack.cs" t="Include" />
+            <e p="EnemySpawner.cs" t="Include" />
+            <e p="GruntScript.cs" t="Include" />
+            <e p="LaserTurret.cs" t="Include" />
+          </e>
+          <e p="Player" t="Include">
+            <e p="AttackPhase.cs" t="Include" />
+            <e p="ChoosePhase.cs" t="Include" />
+            <e p="MovePhase.cs" t="Include" />
             <e p="PlayerController.cs" t="Include" />
-            <e p="PlayerPhases.cs" t="Include" />
+            <e p="PlayerPhase.cs" t="Include" />
           </e>
           <e p="UI" t="Include">
             <e p="CameraFollow.cs" t="Include" />
