--- Assets/Scripts/Creatures.DotSettings.user.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4efaaa1ccca8a4ddcb1902c1f9d10de7
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
