+++ Assets/Scripts/Player/AttackPhase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4202cc299ea7d462085c35a2e4b9de9d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
