+++ Assets/Scripts/Player/ChoosePhase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d03bc05ac5f345c1b04022582ce6ecd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
