+++ Assets/Scripts/Player/MovePhase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ed4971291e6c428f906770799e91592
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
